+++ CHANGELOG.md
@@ -12,6 +12,8 @@ New features:
 * `foreign_object.name = value` will now call `Interoplibrary#writeMember("name", value)` instead of `invokeMember("name=", value)`.
 * Always show the Ruby core library files in backtraces (#1414).
 * The Java stacktrace is now shown when sending SIGQUIT to the process, also on TruffleRuby Native, see [Debugging](doc/user/debugging.md) for details (#2041).
+* Calls to foreign objects with a block argument will now pass the block as the last argument.
+* `foreign.name` will now use `invokeMember` if invocable and if not use `readMember`, see `doc/contrib/interop_implicit_api.md` for details.
 
 Bug fixes: